Split test preparation into its own mojo (#327)
